+++ src/consumer/bit-config/component-overrides.js
@@ -1,8 +1,12 @@
 // @flow
 import R from 'ramda';
-import * as RA from 'ramda-adjunct';
 import ComponentBitConfig from './component-bit-config';
-import { MANUALLY_REMOVE_DEPENDENCY } from '../../constants';
+import {
+  MANUALLY_REMOVE_DEPENDENCY,
+  MANUALLY_ADD_DEPENDENCY,
+  OVERRIDE_FILE_PREFIX,
+  OVERRIDE_COMPONENT_PREFIX
+} from '../../constants';
 import type { ConsumerOverridesOfComponent } from './consumer-overrides';
 import { dependenciesFields } from './consumer-overrides';
 
@@ -22,45 +26,40 @@ export default class ComponentOverrides {
   /**
    * overrides of component can be determined by three different sources.
    * 1. component-config. (bit.json/package.json of the component itself).
-   *    authored don't have it, most imported have it, unless they choose not to write package.json/bit.json.
+   *    authored normally don't have it, most imported have it, unless they choose not to write package.json/bit.json.
    * 2. consumer-config. (bit.json/package.json of the consumer when it has overrides of the component).
    * 3. model. (when the component is tagged, the overrides data is saved into the model).
    *
    * the strategy of loading them is simple:
    * if the component config is written to the filesystem, use it (#1).
    * if the component config is not written, it can be for two reasons:
-   * a) it's author. in this case, the config is written into consumer-config (if not exist) on import.
-   * b) it's imported when user chose not to write package.json nor bit.json. in that case, use
+   * a) it's imported and the user chose not to write package.json nor bit.json. in this case, use
    * component from the model.
-   * once you have the config, merge it with the consumer-config.
+   * b) it's author. by default, the config is written into consumer-config (if not exist) on import.
+   * which, in this case, use only consumer-config.
+   * an exception is when an author runs `eject-conf` command to explicitly write the config, then,
+   * use the component-config.
    */
-  static load(
+  static loadFromConsumer(
     overridesFromConsumer: ?ConsumerOverridesOfComponent,
     overridesFromModel: ?ComponentOverridesData,
     componentBitConfig: ?ComponentBitConfig,
-    isAuthor?: boolean = false
+    isAuthor: boolean
   ): ComponentOverrides {
-    const getOverridesFromComponent = () => {
-      if (isAuthor) return null;
-      return componentBitConfig && componentBitConfig.componentHasWrittenConfig
-        ? componentBitConfig.overrides
-        : overridesFromModel;
-    };
-    const overridesFromComponent = getOverridesFromComponent();
-    const overrides = this.mergeConsumerAndComponentOverrides(overridesFromComponent, overridesFromConsumer);
-    return new ComponentOverrides(overrides, overridesFromConsumer);
+    if (componentBitConfig && componentBitConfig.componentHasWrittenConfig) {
+      // $FlowFixMe
+      return new ComponentOverrides(componentBitConfig.overrides, overridesFromConsumer);
+    }
+    if (!isAuthor) {
+      return ComponentOverrides.loadFromScope(overridesFromModel);
+    }
+    return new ComponentOverrides(overridesFromConsumer, overridesFromConsumer);
   }
-  static mergeConsumerAndComponentOverrides(
-    overridesFromComponent: ?ComponentOverridesData,
-    overridesFromConsumer: ?ConsumerOverridesOfComponent
-  ): ?Object {
-    if (RA.isNilOrEmpty(overridesFromComponent) && RA.isNilOrEmpty(overridesFromConsumer)) return null;
-    // at least one overrides is defined
-    if (RA.isNilOrEmpty(overridesFromConsumer)) return overridesFromComponent;
-    if (RA.isNilOrEmpty(overridesFromComponent)) return overridesFromConsumer;
-    // we don't propagate by default. so if overrides from component is found, just use it.
-    return overridesFromComponent;
+  static loadFromScope(overridesFromModel: ?ComponentOverridesData = {}) {
+    // $FlowFixMe
+    return new ComponentOverrides(overridesFromModel, {});
   }
+
   static componentOverridesDataFields() {
     return dependenciesFields;
   }
@@ -69,23 +68,124 @@ export default class ComponentOverrides {
     const isDependencyField = (val, field) => fields.includes(field);
     return R.pickBy(isDependencyField, this.overrides);
   }
-  getAllDependenciesOverrides(): Object {
-    return Object.assign(
+  getComponentDependenciesWithVersion(): Object {
+    const allDeps = Object.assign(
       {},
       this.overrides.dependencies,
       this.overrides.devDependencies,
       this.overrides.peerDependencies
     );
+    return this._filterForComponentWithValidVersion(allDeps);
   }
-  getAllDependenciesOverridesFromConsumer(): Object {
-    return Object.assign(
+  getComponentDependenciesWithVersionFromConsumer(): Object {
+    const allDeps = Object.assign(
       {},
       this.overridesFromConsumer.dependencies,
       this.overridesFromConsumer.devDependencies,
       this.overridesFromConsumer.peerDependencies
     );
+    return this._filterForComponentWithValidVersion(allDeps);
   }
-  getIgnored(field: string) {
+  _filterForComponentWithValidVersion(deps: Object): Object {
+    return Object.keys(deps).reduce((acc, current) => {
+      if (this._isValidVersion(deps[current]) && current.startsWith(OVERRIDE_COMPONENT_PREFIX)) {
+        const component = current.replace(OVERRIDE_COMPONENT_PREFIX, '');
+        acc[component] = deps[current];
+      }
+      return acc;
+    }, {});
+  }
+  _isValidVersion(ver: string) {
+    return ver !== MANUALLY_ADD_DEPENDENCY && ver !== MANUALLY_REMOVE_DEPENDENCY;
+  }
+  getIgnored(field: string): string[] {
     return R.keys(R.filter(dep => dep === MANUALLY_REMOVE_DEPENDENCY, this.overrides[field] || {}));
+  }
+  getIgnoredFiles(field: string): string[] {
+    const ignoredRules = this.getIgnored(field);
+    return ignoredRules
+      .filter(rule => rule.startsWith(OVERRIDE_FILE_PREFIX))
+      .map(rule => rule.replace(OVERRIDE_FILE_PREFIX, ''));
+  }
+  getIgnoredComponents(field: string): string[] {
+    const ignoredRules = this.getIgnored(field);
+    return R.flatten(
+      ignoredRules
+        .filter(rule => rule.startsWith(OVERRIDE_COMPONENT_PREFIX))
+        .map(rule => rule.replace(OVERRIDE_COMPONENT_PREFIX, ''))
+        .map(idStr => [idStr, ...this._getComponentNamesFromPackages(idStr)])
+    );
+  }
+
+  /**
+   * it is possible that a user added the component into the overrides as a package.
+   * e.g. `@bit/david.utils.is-string` instead of `@bit/david.utils/is-string`
+   * or, if not using bit.dev, `@bit/utils.is-string` instead of `@bit/utils/is-string`
+   */
+  _getComponentNamesFromPackages(idStr: string): string[] {
+    const idSplitByDot = idStr.split('.');
+    const numberOfDots = idSplitByDot.length - 1;
+    if (numberOfDots === 0) return []; // nothing to do. it wasn't entered as a package
+    const localScopeComponent = idSplitByDot.join('/'); // convert all dots to slashes
+    if (numberOfDots === 1) {
+      // it can't be from bit.dev, it must be locally
+      return [localScopeComponent];
+    }
+    // there are two dots or more. it can be from bit.dev and it can be locally
+    // for a remoteScopeComponent, leave the first dot and convert only the rest to a slash
+    const remoteScopeComponent = `${R.head(idSplitByDot)}.${R.tail(idSplitByDot).join('/')}`;
+    return [localScopeComponent, remoteScopeComponent];
+  }
+  getIgnoredPackages(field: string): string[] {
+    const ignoredRules = this.getIgnored(field);
+    return ignoredRules.filter(
+      rule => !rule.startsWith(OVERRIDE_FILE_PREFIX) && !rule.startsWith(OVERRIDE_COMPONENT_PREFIX)
+    );
+  }
+  stripOriginallySharedDir(sharedDir: ?string) {
+    if (!sharedDir) return;
+    dependenciesFields.forEach((field) => {
+      if (!this.overrides[field]) return;
+      Object.keys(this.overrides[field]).forEach((rule) => {
+        if (!rule.startsWith(OVERRIDE_FILE_PREFIX)) return;
+        const fileWithSharedDir = rule.replace(OVERRIDE_FILE_PREFIX, '');
+        // $FlowFixMe we made sure that sharedDir is not empty
+        const fileWithoutSharedDir = fileWithSharedDir.replace(`${sharedDir}/`, '');
+        // $FlowFixMe
+        const value = this.overrides[field][rule];
+        // $FlowFixMe
+        delete this.overrides[field][rule];
+        // $FlowFixMe
+        this.overrides[field][`${OVERRIDE_FILE_PREFIX}${fileWithoutSharedDir}`] = value;
+      });
+    });
+  }
+  addOriginallySharedDir(sharedDir: ?string) {
+    if (!sharedDir) return;
+    dependenciesFields.forEach((field) => {
+      if (!this.overrides[field]) return;
+      Object.keys(this.overrides[field]).forEach((rule) => {
+        if (!rule.startsWith(OVERRIDE_FILE_PREFIX)) return;
+        const fileWithoutSharedDir = rule.replace(OVERRIDE_FILE_PREFIX, '');
+        // $FlowFixMe
+        const fileWithSharedDir = `${sharedDir}/${fileWithoutSharedDir}`;
+        // $FlowFixMe
+        const value = this.overrides[field][rule];
+        // $FlowFixMe
+        delete this.overrides[field][rule];
+        // $FlowFixMe
+        this.overrides[field][`${OVERRIDE_FILE_PREFIX}${fileWithSharedDir}`] = value;
+      });
+    });
+  }
+  static getAllFilesPaths(overrides: Object): string[] {
+    if (!overrides) return [];
+    const allDeps = Object.assign({}, overrides.dependencies, overrides.devDependencies, overrides.peerDependencies);
+    return Object.keys(allDeps)
+      .filter(rule => rule.startsWith(OVERRIDE_FILE_PREFIX))
+      .map(rule => rule.replace(OVERRIDE_FILE_PREFIX, ''));
+  }
+  clone(): ComponentOverrides {
+    return new ComponentOverrides(R.clone(this.overrides), R.clone(this.overridesFromConsumer));
   }
 }